Move #640 concurrency tests to test_schema_operations.py and trim comments
PolymorphicMultiObjectConcurrencyTestCase exercises through-model
registration during polymorphic multiobject field creation (a schema
operation), not deletion logic -- it only lived in test_deletion.py because
the investigation started from the bug's delete-time symptom. Moved it
next to the other schema-creation/registry tests it actually belongs with.

Also trimmed the docstrings and inline comments, which had grown into
multi-paragraph explanations restating the same points -- cut to the
essential why (what's already locked, what isn't, and why the fixed case
times out rather than deadlocking) without re-deriving the whole
investigation inline.
