+++ netbox_custom_objects/tests/test_deletion.py
@@ -6,16 +6,12 @@ setup and teardown are not wrapped in Django's per-test rollback transaction.  T
 lets us verify table-level changes and FK SET NULL/CASCADE/PROTECT behaviour that
 cannot be observed inside a rolled-back savepoint.
 """
-import threading
-
 from django.apps import apps as django_apps
 from django.db import connection
 from django.db.utils import IntegrityError
 from django.test import TransactionTestCase
 
-from core.models import ObjectType
 from dcim.models import Device, DeviceRole, DeviceType, Manufacturer, Site
-from extras.choices import CustomFieldTypeChoices
 from netbox_custom_objects.choices import ObjectFieldOnDeleteChoices
 from netbox_custom_objects.constants import APP_LABEL
 from netbox_custom_objects.models import CustomObjectType, CustomObjectTypeField
@@ -802,253 +798,3 @@ class DeletionTestCase(TransactionCleanupMixin, CustomObjectsTestCase, Transacti
         obj_target = target_model.objects.create(name='Target Object')
         obj_source = source_model.objects.create(name='Source Object', ref_target=obj_target)
         self.assertEqual(obj_source.ref_target, obj_target)
-
-
-class PolymorphicMultiObjectConcurrencyTestCase(TransactionCleanupMixin, CustomObjectsTestCase, TransactionTestCase):
-    """
-    Regression tests for issue #640: concurrent regeneration of a COT with a
-    polymorphic multiobject field could register two competing through-model
-    classes for the same name, leaving a stale "source" FK reference that later
-    surfaced as "ValueError: Cannot query 'X': Must be 'TableYModel' instance."
-    (the same symptom class as #477/#483) or a RecursionError, depending on
-    which thread's registration "won". Confirmed live under a real multi-threaded
-    gunicorn worker; reproducing the exact race deterministically in-process isn't
-    feasible, so this drives many genuinely concurrent get_model() calls through
-    the same code path and asserts the result is always self-consistent.
-    """
-
-    def setUp(self):
-        super().setUp()
-        self.site_ot = ObjectType.objects.get_for_model(Site)
-
-    def test_field_creation_racing_concurrent_readers_yields_consistent_through_model(self):
-        """
-        Racing bare get_model() calls against each other (no I/O between the
-        LookupError check and the register_model() write) rarely lands in the
-        actual race window -- the critical section is nearly pure Python with no
-        GIL-releasing I/O, so threads rarely get preempted inside it. What
-        reproduced this reliably live (issue #640) was racing the polymorphic
-        field's *creation* -- which does real, GIL-releasing DB I/O across several
-        statements (INSERT the field row, then several more for
-        related_object_types.set()) -- against other threads continuously calling
-        get_model(), which is exactly the shape of "one request creates a field
-        while other requests are rendering unrelated pages" in production.
-        """
-        cot = self.create_simple_custom_object_type(name='polyrace', slug='poly-race')
-
-        stop = threading.Event()
-        reader_errors = []
-        reader_errors_lock = threading.Lock()
-
-        def reader():
-            while not stop.is_set():
-                try:
-                    CustomObjectType.objects.get(pk=cot.pk).get_model(no_cache=True)
-                except Exception as e:  # noqa: BLE001 - captured for the assertion below
-                    with reader_errors_lock:
-                        reader_errors.append(e)
-                finally:
-                    connection.close()
-
-        n_readers = 12
-        readers = [threading.Thread(target=reader) for _ in range(n_readers)]
-        for t in readers:
-            t.start()
-
-        try:
-            field = self.create_custom_object_type_field(
-                cot,
-                name='depends_on',
-                label='Depends On',
-                type='multiobject',
-                is_polymorphic=True,
-            )
-            field.related_object_types.set([self.site_ot])
-        finally:
-            stop.set()
-            for t in readers:
-                t.join()
-
-        self.assertEqual(
-            reader_errors, [],
-            "concurrent get_model() calls must not raise while a polymorphic "
-            "multiobject field is being created",
-        )
-
-        # Self-consistency: whatever get_model() now returns must be the same class
-        # the registered through model's "source" FK actually points at -- a
-        # mismatch here is exactly the #477/#483-class staleness this guards against.
-        final_model = CustomObjectType.objects.get(pk=cot.pk).get_model()
-        through_model = django_apps.get_model(APP_LABEL, field.through_model_name)
-        source_field = through_model._meta.get_field('source')
-        self.assertIs(
-            source_field.remote_field.model, final_model,
-            "the registered through model's source FK must point at the model class "
-            "get_model() currently returns, not an orphaned duplicate from a losing thread",
-        )
-
-        # And that consistency must actually be usable: creating an instance and
-        # relating it through the polymorphic field, then deleting it, must not raise
-        # the #477/#483-class ValueError.
-        obj = final_model.objects.create(name='Instance 1')
-        obj.depends_on.set([Site.objects.create(name='Race Site', slug='race-site')])
-        obj.delete()  # Must not raise ValueError: "Cannot query ...: Must be ... instance."
-
-    def test_forced_registration_interleaving_stays_consistent(self):
-        """
-        Deterministic version of the race, forced via mocking rather than relying
-        on real thread-scheduling luck (which the test above showed rarely lands
-        inside the narrow window this actually depends on).
-
-        The real #640 race is NOT two get_model() readers colliding with each
-        other -- CustomObjectType.get_model() already wraps its whole call to
-        _after_model_generation() in CustomObjectType._global_lock, so two
-        concurrent readers regenerating the same COT are already fully
-        serialized there, with or without any change targeting that function.
-
-        The actual gap is on the *writer* side:
-        CustomFieldType.create_polymorphic_m2m_table() (called exactly once,
-        from CustomObjectTypeField.save(), when a new polymorphic multiobject
-        field is first created) builds a fresh through-model class, lets
-        Django's ModelBase metaclass register it, and only *afterwards* points
-        its "source" FK at the caller's model -- all with no lock at all. A
-        concurrent reader's get_model(no_cache=True) -- lock-protected only on
-        its own side -- can run in that exact window: it finds the writer's
-        through model already registered (via the metaclass) and immediately
-        repoints "source" at ITS OWN freshly-regenerated model class. Whichever
-        of the two threads mutates "source" last, and whichever one's
-        get_model() call caches its own model last, aren't guaranteed to be
-        the same thread -- so the registered through's "source" FK and
-        whatever get_model() now returns can end up pointing at two different
-        (if table-equivalent) Python classes. Confirmed live under concurrent
-        load: intermittent "ValueError: Cannot query 'X': Must be
-        'TableYModel' instance." and RecursionError.
-
-        This test forces exactly that interleaving: thread "W" plays the
-        writer (calling create_polymorphic_m2m_table() directly, as
-        CustomObjectTypeField.save() would), thread "R" plays the reader
-        (get_model(no_cache=True)). A mocked apps.register_model() hook pauses
-        W immediately after its metaclass-driven registration -- but *before*
-        W repoints "source" at its own model -- and only resumes W once R has
-        had its chance to run. With the #640 fix, W's entire
-        create_polymorphic_m2m_table() body (including that registration) now
-        runs under CustomObjectType._global_lock, so R can't even start its
-        own lock-protected check until W's whole turn -- pause included --
-        is over; the rendezvous below simply times out and W proceeds alone,
-        R correctly reuses W's finished result afterward. Without the fix, R
-        genuinely runs inside the pause and the two threads' "source"
-        FK/get_model() cache writes land in different orders, reliably
-        producing the mismatch this test asserts against.
-        """
-        from unittest.mock import patch
-
-        from netbox_custom_objects.field_types import FIELD_TYPE_CLASS
-
-        cot = self.create_simple_custom_object_type(name='polyforce', slug='poly-force')
-        field = self.create_custom_object_type_field(
-            cot,
-            name='depends_on',
-            label='Depends On',
-            type='multiobject',
-            is_polymorphic=True,
-        )
-        field.related_object_types.set([self.site_ot])
-
-        # The real table/through model already exist (created for real by
-        # create_custom_object_type_field() above via the normal save() path).
-        # Force the through model back to "unregistered" so a direct call to
-        # create_polymorphic_m2m_table() -- simulating field creation racing a
-        # concurrent reader, as CustomObjectTypeField.save() would trigger --
-        # takes the same "build fresh, register, then repoint source" path a
-        # brand-new field's first save would. The physical table is left
-        # alone; create_polymorphic_m2m_table()'s own idempotency check will
-        # see it already exists and skip re-issuing the DDL.
-        writer_model = CustomObjectType.objects.get(pk=cot.pk).get_model()
-        CustomObjectType.clear_model_cache()
-        model_name_lower = field.through_model_name.lower()
-        del django_apps.all_models[APP_LABEL][model_name_lower]
-        django_apps.clear_cache()
-
-        real_register_model = django_apps.register_model
-        reader_may_proceed = threading.Event()
-        reader_done = threading.Event()
-        gated = set()
-
-        def ordered_register_model(app_label, model):
-            # Only intercept the through model under test; everything else
-            # (e.g. the reader's own source-model registration) is untouched.
-            if app_label != APP_LABEL or model.__name__ != field.through_model_name:
-                return real_register_model(app_label, model)
-
-            # Only the FIRST call matters -- Django's ModelBase metaclass
-            # registers the model as soon as generate_model() builds it
-            # (inside get_polymorphic_through_model()); this is that call.
-            if 'seen' in gated:
-                return real_register_model(app_label, model)
-            gated.add('seen')
-
-            result = real_register_model(app_label, model)
-            # The through model is now registered but W (the writer) hasn't
-            # yet repointed its "source" FK at writer_model -- give R (the
-            # reader) a chance to run right here. With the #640 fix, W is
-            # holding CustomObjectType._global_lock for this whole call, so R
-            # can't have even started its own check yet; this just times out
-            # and W proceeds immediately.
-            reader_may_proceed.set()
-            reader_done.wait(timeout=2)
-            return result
-
-        writer_result = {}
-
-        def run_writer():
-            threading.current_thread().name = 'W'
-            field_type = FIELD_TYPE_CLASS[CustomFieldTypeChoices.TYPE_MULTIOBJECT]()
-            try:
-                with connection.schema_editor() as schema_editor:
-                    field_type.create_polymorphic_m2m_table(field, writer_model, schema_editor)
-            except Exception as e:  # noqa: BLE001 - surfaced via the assertion below
-                writer_result['error'] = e
-            finally:
-                connection.close()
-
-        reader_result = {}
-
-        def run_reader():
-            threading.current_thread().name = 'R'
-            reader_may_proceed.wait(timeout=5)
-            try:
-                reader_result['model'] = CustomObjectType.objects.get(pk=cot.pk).get_model(no_cache=True)
-            except Exception as e:  # noqa: BLE001 - surfaced via the assertion below
-                reader_result['error'] = e
-            finally:
-                reader_done.set()
-                connection.close()
-
-        with patch.object(django_apps, 'register_model', side_effect=ordered_register_model):
-            t_w = threading.Thread(target=run_writer, name='W')
-            t_r = threading.Thread(target=run_reader, name='R')
-            t_w.start()
-            t_r.start()
-            t_w.join(timeout=10)
-            t_r.join(timeout=10)
-
-        self.assertNotIn('error', writer_result, f"writer raised: {writer_result.get('error')!r}")
-        self.assertNotIn('error', reader_result, f"reader raised: {reader_result.get('error')!r}")
-
-        # The invariant the fix establishes: whichever model get_model() now
-        # returns must be the one the registered through model's "source" FK
-        # actually points at. Without the #640 fix, this forced interleaving
-        # reliably produces a mismatch (reader's model cached, writer's model
-        # left on the through's "source" FK, or vice versa) every time.
-        final_model = CustomObjectType.objects.get(pk=cot.pk).get_model()
-        through_model = django_apps.get_model(APP_LABEL, field.through_model_name)
-        source_field = through_model._meta.get_field('source')
-        self.assertIs(
-            source_field.remote_field.model, final_model,
-            "the registered through model's source FK must point at the model class "
-            "get_model() currently returns -- a mismatch here is issue #640",
-        )
-
-        obj = final_model.objects.create(name='Instance 1')
-        obj.depends_on.set([Site.objects.create(name='Force Site', slug='force-site')])
-        obj.delete()  # Must not raise ValueError: "Cannot query ...: Must be ... instance."
